+++ Config.php
@@ -40,7 +40,8 @@ class Config
         'user_access_attribute_server_specification_delimiter' => ';',
         'user_access_attribute_server_separator' => ':',
         'instance_name' => '',
-        'ldap_network_timeout' => Client::DEFAULT_TIMEOUT_SECS
+        'ldap_network_timeout' => Client::DEFAULT_TIMEOUT_SECS,
+        'enable_password_confirmation' => 0
     );
 
     // for backwards compatibility
@@ -194,7 +195,7 @@ class Config
 
     public static function getServerConfig($server)
     {
-        $configName = 'LoginLdap_' . $server;
+        $configName = 'LoginLdap_' . preg_replace('/[^a-zA-Z0-9_-]/', '', $server);
         return PiwikConfig::getInstance()->__get($configName);
     }